Make sure we don't get the other branch of the hyperbolae
This improves the plots for the proplyd fits. We still have a problem
with low-eccentricity ellipses. We need to find a way to force the
focus to be close to the data points. We could calulate the distance
of the two foci and always switch to the one that is closest.