Fix formatting of notebooks that black had messed up
Turns out that although black knows not to remove semicolons from
ipynb files, it does not know not to remove them from python pies that
are synched with those notebooks. I have fixed this by hand for now
and turned off the auto-black formatting in my emacs config. A better
fix would be to switch to ruff for all formatting and linting tasks,
since that has more options for configuration.